Added location picker in search and refactoring sighting table and sighting filter in seperate files
