+++ ui/src/components/formComponents/SightingsFilter/SightingFilter.scss
@@ -1,0 +1,118 @@
+@use "sass:color";
+@use '@/styles/constants';
+
+// Desktop styling
+img{
+    width: 50%;
+    height: auto;
+}
+
+table {
+    width: 100%;
+    border: 1px solid;
+}
+
+.table-container {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    margin: 40px auto;
+    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+    background-color: #fff;
+    border-radius: constants.$border-radius;
+    font-size: constants.$font-size-normal;
+    font-family: constants.$font-family-normal;
+    overflow: hidden;
+}
+ 
+.table-header, .table-row {
+    display: flex;
+    justify-content: space-between;
+    padding: constants.$padding-small constants.$padding-large;
+    border-bottom: 1px solid #ddd;
+}
+ 
+.table-header {
+    background-color: #5d626f;
+    color: #fff;
+    font-weight: 600;
+    text-transform: uppercase;
+}
+ 
+.table-row:nth-child(even) {
+    background-color: #f9f9f9;
+}
+ 
+.table-row:hover {
+    background-color: #d6e9f5;
+}
+.table-cell {
+    flex: 1;
+    text-align: left;
+}
+ 
+.table-cell.status {
+    flex: 1;
+    text-align: center;
+}
+
+.pagination-links {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    gap: 16px;
+  }
+  
+  a:hover {
+    background-color: #ddd;
+    color: black;
+  }
+  
+  .previous {
+    background-color: #f1f1f1;
+    color: black;
+  }
+  
+  .next {
+    background-color: #04AA6D;
+    color: white;
+  }
+  
+  .filterContainer {
+    display: flex;
+    flex-direction: row;
+    gap: 16px;
+  }
+
+// Mobile styling
+@media (max-width: constants.$tablet-min-width) {
+    .ul-container {
+        display: flex;
+        flex-direction: column;
+        width: 100%;
+        margin: 40px auto;
+        box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+        background-color: #fff;
+        border-radius: constants.$border-radius;
+        overflow: hidden;
+        margin-right: auto;
+        padding: constants.$padding-default;
+        gap: 10px;
+    }
+    
+    .li-item {
+        list-style-type: none;
+        border-radius: constants.$border-radius;;
+        border: 2px solid constants.$background-navigation;;
+        align-content: flex-start;
+        padding: constants.$padding-default;
+        font-size: constants.$font-size-normal;
+        font-family: constants.$font-family-normal;
+    }
+    
+    img {
+        width: 90%;
+        height: auto;
+        border-radius: constants.$border-radius;
+    }   
+}
